requirements.txt
fastapi>=0.110,<0.117
uvicorn[standard]>=0.27,<0.36
jinja2>=3.1,<3.2
python-multipart>=0.0.9,<0.0.21
python-dotenv>=1,<2
sqlalchemy>=2.0,<2.1
psycopg[binary]>=3.1,<3.3
httpx>=0.27,<0.29
itsdangerous>=2.2,<2.3
